[IMP] make the bracket selector always find the first opening bracket
Purpose
=======
Make the bracket selector always find the first opening bracket.

It has the advantage to give more option, and to be consistent when
doing back and forth (doing the next command is the go back to the
state before the next command, in the context of nested bracket).

So now, if we are inside a bracket, we need to do "select previous"
to select the parent (before both command will work).